# Fix backend benchmarks + a micro-optimization