@greatroar 2020-06-14
Fair enough, reinstated the checks with StopTimer/StartTimer around them.