@MichaelEischer 2020-06-16
@greatroar There's a missing call to restart the timer again (the benchmark actually runs into a timeout). Can you please add that? Afterwards this PR is ready for merging.